added basic crawler and example
