+++ docs/source/crawlers.rst
@@ -1,0 +1,5 @@
+Crawlers
+========
+
+.. autoclass:: rasp.crawlers.base.DefaultBoundedCrawler
+    :members:
