+++ docs/source/index.rst
@@ -33,6 +33,7 @@ yet, so install with:
 
    webpage
    engines
+   crawlers
