+++ rasp/__init__.py
@@ -1,8 +1,7 @@
-from rasp.base import DefaultEngine
-from rasp.tor_engine import TorEngine
+from rasp.engines import *
+from rasp.crawlers import *
 
-__all__ = [
-    'DefaultEngine',
-    'SeleniumEngine',
-    'TorEngine'
-]
+from rasp import engines
+from rasp import crawlers
+
+__all__ = engines.__all__ + crawlers.__all__
