+++ rasp/crawlers/__init__.py
@@ -1,0 +1,5 @@
+from rasp.crawlers.base import DefaultBoundedCrawler
+
+__all__ = [
+    "DefaultBoundedCrawler"
+]
