+++ rasp/engines/__init__.py
@@ -1,0 +1,9 @@
+from rasp.engines.base import DefaultEngine
+from rasp.engines.selenium_engine import SeleniumEngine
+from rasp.engines.tor_engine import TorEngine
+
+__all__ = [
+    'DefaultEngine',
+    'SeleniumEngine',
+    'TorEngine'
+]
